Move time-related stuff from column to fitstime.
